requirements.txt
requests~=2.32.3
psycopg2~=2.9.10
uvicorn~=0.34.2
python-dotenv~=1.1.0
fastapi~=0.115.12
mcp~=1.9.0
asyncpg~=0.30.0
httpx~=0.28.1